Add module import routine
Introduce module import routine into runtime, which takes either
absolute or relative path, converts that into a path which points to a
file in the file system, compiles and finally runs that file. If
everything went smoothly, words from that file are being exported into
dictionary of current execution context.