Different approach on arrays
Instead of always relying on arrays which are only wrappers to std::vector,
introduce abstract array class with multiple different implementations on it.

This in theory could reduce memory usage and binary size, especially with
scripts that combine arrays with each other. Somewhat experimental idea that
could be expanded also to string values.