Modify compiler to use Unicode strings
Instead of having to rely on byte strings that are only assumed to be
UTF-8 encoded, modify the compiler so that it only accepts Unicode
strings. This means that that the decoding must be done by whatever is
calling the compilation, so that the compiler no longer has to worry
about charcter encoding.

This modification is particularily done with WebAssembly target in mind.
Emscripten seems to use UTF-16 as character encoding and is not capable
of handling UTF-8 strings.