Rename plorth-cli executable to just plorth
The real reason why name of Plorth interpreter is currently `plorth-cli`
is because I want to separate the interpreter library and executable.
CMake cannot handle targets with the same name, so I cannot have both
library and executable called `plorth`. However, you can override the
executable name with this simple trick which gives us both library and
executable with the same name.