Include source code position in symbols
Modify `symbol` class so that it now contains optional position in the
source code where the symbol was encountered. Modify the compiler so
that it tracks position of symbols in the source code, unless symbol
caching has been enabled. Also fixes a bug which actually enables
caching of compiled symbols, which wasn't previously done for some
reason.