Determine target automatically
Instead of having to manually select whether we want to compile CLI
version or WebAssembly version of the interpreter, determine it
automatically by looking for `EMSCRIPTEN` environment variable.

It doesn't automatically turn on 32bit integers, disable modules and so
on yet. Those need to be still configured manually for each build.