Make dictionary entries "editable" in GUI
Modify the dictionary display component so that it signals when an entry
has been double clicked by the user. Then attach to this signal in the
main window, so that source code of the selected word will be placed
into the line editor, thus making the dictionary entries kinda editable.