Do not use formatted output
I/O streams in C++ are nice, but the formatted output on them is
obviously more slow than using `std::fwrite` etc. directly, which is why
the I/O streams should not be used in standard output of the
interpreter. Also ditch usage of iterators in `utf8_encode` function
which seems to improve overall performance.