Replace linenoise with UTF-8 compatible fork
Linenoise library used to implement line editing in Plorth REPL is
excellent minimalistic implementation that is easy to embed into Plorth
CLI, but it's written in C and is missing UTF-8 support. Because of
this, I replaced the stock linenoise with my own fork of it which has
been largely rewritten in C++11, includes UTF-8 support and works
directly with Unicode strings.